+++ OSVRUnreal/Plugins/OSVR/Source/OSVRInput/Private/OSVRInputPrivatePCH.h
@@ -1,0 +1,10 @@
+#include "Core.h"
+#include "CoreUObject.h"
+
+#include "Engine.h"
+
+#include <osvr/ClientKit/ContextC.h>
+#include <osvr/ClientKit/InterfaceC.h>
+#include <osvr/ClientKit/InterfaceCallbackC.h>
+#include <osvr/ClientKit/InterfaceStateC.h>
+#include <osvr/ClientKit/ParametersC.h>
